reajust tree

+++ lib/fichier.h
@@ -4,5 +4,5 @@
 #include "../lib/texte.h"
 
 Texte lire_fichier(char *nom_fichier);
-int taille_fichier(FILE * fichier);
+long taille_fichier(FILE * fichier);
 int nombre_lignes(FILE *fichier);
